refactor(dev): vanilla extract namespace
