+++ packages/remix-dev/compiler/css/compiler.ts
@@ -20,7 +20,7 @@ import { urlImportsPlugin } from "../plugins/urlImportsPlugin";
 // import { cssBundleUpdatePlugin } from "../plugins/cssBundleUpdatePlugin";
 import { cssModulesPlugin } from "../plugins/cssModulesPlugin";
 import { cssSideEffectImportsPlugin } from "../plugins/cssSideEffectImportsPlugin";
-import { vanillaExtractPlugin } from "../plugins/vanillaExtractPlugin";
+import { vanillaExtractPlugin } from "../plugins/vanillaExtract";
 import {
   cssBundleEntryModulePlugin,
   cssBundleEntryModuleId,
